requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
cffi==1.14.4
channels==3.0.3
channels-redis
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
hyperlink==20.0.1
idna==3.1
incremental==17.5.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.0
pyOpenSSL==20.0.1
pytz==2020.5
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
Twisted==20.3.0
txaio==20.12.1
whitenoise==5.2.0
zope.interface==5.2.0
